requirements.txt
# Core dependencies
langchain>=1.0.8
langgraph>=1.0.3
chromadb>=1.3.5
langchain-google-genai>=3.1.0
langchain-chroma>=1.0.0
pydantic>=2.12.4
python-dotenv>=1.2.1
rank_bm25>=0.2.2
tqdm>=4.67.1
langchain-text-splitters>=1.0.0

# Development tools
black>=24.0.0
isort>=5.13.0
ruff>=0.1.0
mypy>=1.8.0
pytest>=7.4.0
pytest-cov>=4.1.0
